Move the route decorators into one module
In preparation for sharing more logic between the route decorators,
move them into a single module.
